pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
slowapi = "^0.1.9"
limits = "^3.10.1"
starlette = "^0.37.2"
apscheduler = "^3.10.4"
requests = "^2.31.0"
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
python-dotenv = "^1.0.1"
opentelemetry-instrumentation-requests = "^0.45b0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
xds-protos = "^1.62.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"